feat(shared-modules-refactor): Sub-PR 3 — extraer <JuntaDetailModule>
Sigue la convención SM1-SM5 codificada en ADR-011: extrae las ~1346
líneas duplicadas entre `app/rdb/admin/juntas/[id]/page.tsx` (1801) y
`app/dilesa/admin/juntas/[id]/page.tsx` (1817) a un componente
compartido. Pages quedan en 12 líneas cada uno.

## Análisis previo (clasificación de las 606 líneas divergentes)

- **53 líneas cosmético** → resueltas con prop `empresaSlug` (URL de
  retorno se construye internamente).
- **196 líneas lógica divergente legítima** → adoptadas patrones
  superiores de DILESA (decisiones técnicas en planning doc).
- **215 líneas drift accidental** → unificadas a la versión correcta.

## Decisiones técnicas

- **Validación estricta de tareas (opción A)**: el componente shared
  exige `titulo` + `prioridad` + `asignado_a` + `fecha_vence`. Adopta
  el patrón de DILESA — RDB antes solo exigía `titulo`. Schema permite
  NULL pero la disciplina UX evita tareas huérfanas sin dueño.
- **`fecha_vence` + `fecha_compromiso` en sync**: el insert escribe
  ambos campos (mismo patrón que el update ya hacía). Resuelve el
  drift donde RDB escribía solo `fecha_vence` y DILESA solo
  `fecha_compromiso`.
- **`empresaSlug` única prop**: `empresa_id` se lee de `juntaData`
  (ya está en la fila), no hay branding visual específico → la URL
  de retorno es la única diferencia.

## Bugs heredados arreglados oportunísticamente

- **Auto-save de notas no normalizaba URLs en RDB**: persistía signed
  URLs (TTL 6h) que rompían imágenes al expirar. Adoptado
  `normalizeHtmlImagesToPaths()` de DILESA.
- **Editor readiness con race condition en RDB**: hidratación inicial
  dentro de `fetchAll` podía dispararse antes de TipTap listo.
  Adoptado `useEffect` separado de DILESA.
- **`selectedPersonaId` state local innecesario en RDB**: el patrón
  de DILESA usa el `onChange` del Combobox directamente (un click
  menos).
- **Title del checkbox "Enviar a Consejo" decía hardcoded
  `consejo@dilesa.mx` en ambos pages**: drift por copy-paste.
  Reemplazado con mensaje genérico.

## Validación local

- typecheck ✅
- lint ✅ (sin warnings nuevos)
- format:check ✅
- 222 tests ✅

Smoke pendiente en preview deploy:
- /rdb/admin/juntas/<id>: edición, participantes, notas con auto-save,
  tareas con avances, terminar/reenviar/eliminar.
- /dilesa/admin/juntas/<id>: idem.
- Validar especialmente: imágenes embebidas no se rompen al recargar
  (RDB heredera el fix de normalización).
